requirements.txt
numpy==1.16.4
streamlit==0.52.1
pillow
keras
tensorflow==2.0.0b1
